(portage-#25) Don't modify persistent hash contents
The previous implementation of PortageFile.build_line was modifying the
hash representation of a record when constructing lines, so each
successive write of the record would modify it more.

This would only affect package_use and package_keywords since they would
need to modify the record line while the portage_mask and portage_unmask
would directly write :name to disk.