Default package_* to present
Previously if the ensure value was not specified and no other properties
or parameters were given, package_* entries would be treated as
unmanaged. This commit defaults the ensure to present for all package_
types.

This commit is based on 31b53b9738448c2d7844b0c0523b0315e3eb0b4e
It fixes #41 and causes no regression to #45